работа с сервисом Ads

+++ src/main/java/ru/skypro/homework/repository/AdsRepository.java
@@ -1,5 +1,6 @@
 package ru.skypro.homework.repository;
 
+import io.swagger.models.auth.In;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.skypro.homework.entity.Ads;
@@ -8,7 +9,7 @@ import java.util.Collection;
 
 
 @Repository
-public interface AdsRepository extends JpaRepository<Ads, Long> {
+public interface AdsRepository extends JpaRepository<Ads, Integer> {
 
     void deleteAllById(Long adsId);
